refactor: extract a provider-keys hook and resources panel from the ai settings tab
Move provider key/model state and the save/test/remove/pull handlers into
useProviderKeys, and the inline Ollama RAM/VRAM block into OllamaResourcesPanel
(which now fetches its own system resources). index.tsx is the provider-row render.
